+++ src/extension.ts
@@ -1,5 +1,7 @@
 import * as vscode from "vscode";
 import { switchBranches } from "./switch-branches";
+import { deleteStaleBranches } from "./prune-stale-branches";
+import { switchToDefaultBranch } from "./switch-to-default-branch";
 
 export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(
@@ -7,6 +9,18 @@ export function activate(context: vscode.ExtensionContext) {
             "multi-repo-branch-switcher.switchBranches",
             async () => {
                 await switchBranches();
+            }
+        ),
+        vscode.commands.registerCommand(
+            "multi-repo-branch-switcher.deleteStaleBranches",
+            async () => {
+                await deleteStaleBranches();
+            }
+        ),
+        vscode.commands.registerCommand(
+            "multi-repo-branch-switcher.switchToDefaultBranch",
+            async () => {
+                await switchToDefaultBranch();
             }
         )
     );